Minor Quality of Life improvements (#92)
* Standardizes error codes for all commands

* Adds ability to calculate the SHA-256 hash of the recipe parser's contents

* Moves V*_RECIPE_FILE_NAME variables to a common location to prevent duplication of constants

* rattler-bulk-build now requires directories to be passed in

* Moves messaging classes to a more available location. It no longer makes sense as being specific to recipe-parsing
