+++ conda_recipe_manager/parser/types.py
@@ -5,7 +5,7 @@ Description:    Provides public types, type aliases, constants, and small classe
 
 from __future__ import annotations
 
-from enum import StrEnum, auto
+from enum import StrEnum
 from typing import Final
 
 from conda_recipe_manager.parser.enums import SchemaVersion
@@ -22,6 +22,11 @@ NodeValue = Primitives | list[str]
 # The "new" recipe format introduces the concept of a schema version. Presumably the "old" recipe format would be
 # considered "0". When converting to the V1 format, we'll use this constant value.
 CURRENT_RECIPE_SCHEMA_FORMAT: Final[int] = SchemaVersion.V1.value
+
+# Pre-CEP-13 name of the recipe file
+V0_FORMAT_RECIPE_FILE_NAME: Final[str] = "meta.yaml"
+# Required file name for the recipe, specified in CEP-13
+V1_FORMAT_RECIPE_FILE_NAME: Final[str] = "recipe.yaml"
 
 # Indicates how many spaces are in a level of indentation
 TAB_SPACE_COUNT: Final[int] = 2
@@ -112,80 +117,3 @@ class MultilineVariant(StrEnum):
     CARROT = ">"
     CARROT_PLUS = ">+"
     CARROT_MINUS = ">-"
-
-
-class MessageCategory(StrEnum):
-    """
-    Categories to classify `RecipeParser` messages into.
-    """
-
-    EXCEPTION = auto()
-    ERROR = auto()
-    WARNING = auto()
-
-
-class MessageTable:
-    """
-    Stores and tags messages that may come up during `RecipeParser` operations. It is up to the client program to
-    handle the logging of these messages.
-    """
-
-    def __init__(self) -> None:
-        """
-        Constructs an empty message table
-        """
-        self._tbl: dict[MessageCategory, list[str]] = {}
-
-    def add_message(self, category: MessageCategory, message: str) -> None:
-        """
-        Adds a message to the table
-        :param category:
-        :param message:
-        """
-        if category not in self._tbl:
-            self._tbl[category] = []
-        self._tbl[category].append(message)
-
-    def get_messages(self, category: MessageCategory) -> list[str]:
-        """
-        Returns all the messages stored in a given category
-        :param category: Category to target
-        :returns: A list containing all the messages stored in a category.
-        """
-        if category not in self._tbl:
-            return []
-        return self._tbl[category]
-
-    def get_message_count(self, category: MessageCategory) -> int:
-        """
-        Returns how many messages are stored in a given category
-        :param category: Category to target
-        :returns: A list containing all the messages stored in a category.
-        """
-        if category not in self._tbl:
-            return 0
-        return len(self._tbl[category])
-
-    def get_totals_message(self) -> str:
-        """
-        Convenience function that returns a displayable count of the number of warnings and errors contained in the
-        messaging object.
-        :returns: A message indicating the number of errors and warnings that have been accumulated. If there are none,
-                  an empty string is returned.
-        """
-        if not self._tbl:
-            return ""
-
-        def _pluralize(n: int, s: str) -> str:
-            if n == 1:
-                return s
-            return f"{s}s"
-
-        num_errors: Final[int] = 0 if MessageCategory.ERROR not in self._tbl else len(self._tbl[MessageCategory.ERROR])
-        errors: Final[str] = f"{num_errors} " + _pluralize(num_errors, "error")
-        num_warnings: Final[int] = (
-            0 if MessageCategory.WARNING not in self._tbl else len(self._tbl[MessageCategory.WARNING])
-        )
-        warnings: Final[str] = f"{num_warnings} " + _pluralize(num_warnings, "warning")
-
-        return f"{errors} and {warnings} were found."
